add type:module in package.json and change all .mjs to .js
