+++ package.json
@@ -3,8 +3,9 @@
   "version": "0.4.0",
   "description": "A treeshake-ready, functional library of Ethereum utility functions",
   "main": "dist/eth-fun.js",
+  "type": "module",
   "scripts": {
-    "build:node": "esbuild src/index.mjs --outfile=dist/eth-fun.js --sourcemap --bundle --platform=node --target=node14.16.1",
+    "build:node": "esbuild src/index.js --outfile=dist/eth-fun.js --sourcemap --bundle --platform=node --target=node14.16.1",
     "git:add": "git update-index --add dist/eth-fun.js dist/eth-fun.js.map",
     "test": "ava --verbose --timeout 20s"
   },
