+++ src/allFunctions.js
@@ -1,0 +1,9 @@
+// @format
+export default function allFunctions(contracts) {
+  let fns = {};
+  for (let [key, impl] of Object.entries(contracts)) {
+    fns[key] = impl.abi.filter(({ type }) => type === "function");
+  }
+
+  return fns;
+}
