+++ src/constants.js
@@ -1,0 +1,7 @@
+//@format
+const constants = {
+  id: 1,
+  jsonrpc: "2.0"
+};
+
+export default constants;
