--- src/constants.mjs
@@ -1,7 +1,0 @@
-//@format
-const constants = {
-  id: 1,
-  jsonrpc: "2.0"
-};
-
-export default constants;
