+++ src/getStorageLocation.js
@@ -1,0 +1,5 @@
+//@format
+export default function getStorageLocation(contract, soughtLabel) {
+  const { storageLayout } = contract;
+  return storageLayout.storage.find(({ label }) => label === soughtLabel);
+}
