--- src/getStorageLocation.mjs
@@ -1,5 +1,0 @@
-//@format
-export default function getStorageLocation(contract, soughtLabel) {
-  const { storageLayout } = contract;
-  return storageLayout.storage.find(({ label }) => label === soughtLabel);
-}
