--- src/index.mjs
@@ -1,30 +1,0 @@
-// @format
-import allFunctions from "./allFunctions.mjs";
-import getStorageLocation from "./getStorageLocation.mjs";
-import compile from "./compile.mjs";
-import blockNumber from "./blockNumber.mjs";
-import getBlockByNumber from "./getBlockByNumber.mjs";
-import { getStorageAt } from "./getStorageAt.mjs";
-import nodes from "./nodes.mjs";
-import { RPCError } from "./errors.mjs";
-import { call, encodeCallSignature, decodeCallOutput } from "./call.mjs";
-import { toHex } from "./utils.mjs";
-
-const errors = {
-  RPCError
-};
-
-export {
-  allFunctions,
-  compile,
-  blockNumber,
-  getStorageAt,
-  getStorageLocation,
-  nodes,
-  errors,
-  call,
-  encodeCallSignature,
-  decodeCallOutput,
-  toHex,
-  getBlockByNumber,
-};
