--- src/nodes.mjs
@@ -1,45 +1,0 @@
-// @format
-// Source: https://github.com/dmihal/ethereum-nodes/blob/491f610e086bfe0e4eef7d7d143f81a733ad92b9/data/nodes.ts#L15
-// Requirments:
-// - no API key necessary
-const nodes = {
-  mainnet: [
-    {
-      name: "MyCrypto",
-      endpoint: "https://api.mycryptoapi.com/eth",
-      website: "https://mycrypto.com/"
-    },
-    {
-      name: "1inch",
-      endpoint: "https://web3.1inch.exchange/",
-      website: "https://1inch.exchange/"
-    },
-    {
-      name: "Cloudflare",
-      endpoint: "https://cloudflare-eth.com/",
-      website: "https://cloudflare-eth.com/"
-    },
-    {
-      name: "Blockscout",
-      endpoint: "https://mainnet-nethermind.blockscout.com/",
-      website: "https://blockscout.com"
-    },
-    {
-      name: "MyEtherWallet",
-      endpoint: "https://nodes.mewapi.io/rpc/eth",
-      website: "https://myetherwallet.com/"
-    },
-    {
-      name: "LinkPool",
-      endpoint: "https://main-rpc.linkpool.io/",
-      website: "https://linkpool.io/"
-    },
-    {
-      name: "AVADO",
-      endpoint: "https://mainnet.eth.cloud.ava.do/",
-      website: "https://ava.do"
-    }
-  ]
-};
-
-export default nodes;
