+++ src/utils.js
@@ -1,0 +1,4 @@
+// @format
+export function toHex(num) {
+  return `0x${num.toString(16)}`;
+}
