--- src/utils.mjs
@@ -1,4 +1,0 @@
-// @format
-export function toHex(num) {
-  return `0x${num.toString(16)}`;
-}
