+++ test/blockNumber_test.js
@@ -1,0 +1,13 @@
+// @format
+import test from "ava";
+
+import getBlockNumber from "../src/blockNumber.js";
+
+test("getting block number", async t => {
+  const options = {
+    url: "https://cloudflare-eth.com",
+  };
+  const no = await getBlockNumber(options);
+  t.is(typeof no, "string");
+  t.true(no.includes("0x"));
+});
