--- test/index_test.mjs
@@ -1,14 +1,0 @@
-// @format
-import test from "ava";
-
-import {errors, call, encodeCallSignature, decodeCallOutput} from "../src/index.mjs";
-
-test("importing errors", t => {
-  t.truthy(errors.RPCError);
-});
-
-test("importing eth call and utility functions", t => {
-  t.truthy(call);
-  t.truthy(encodeCallSignature);
-  t.truthy(decodeCallOutput);
-});
