--- test/readme_test.mjs
@@ -1,29 +1,0 @@
-// @format
-import { readFileSync } from "fs";
-import path from "path";
-import { fileURLToPath } from "url";
-
-import { matchBlocks, exec } from "test-readme-md";
-import test from "ava";
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-let content = readFileSync("./readme.md").toString();
-content = content.replace(
-  new RegExp("eth-fun", "g"),
-  `file://${path.resolve(__dirname, "../src/index.mjs")}`
-);
-const blocks = matchBlocks("js", content);
-
-for (let block of blocks) {
-  test(`run ${block.split("\n")[0]}`, async t => {
-    try {
-      await exec(block);
-    } catch (err) {
-      console.log(err);
-      t.fail();
-    }
-
-    t.pass();
-  });
-}
